Update tugas fungsi faktorial.cpp

+++ tugas fungsi faktorial.cpp
@@ -1,5 +1,5 @@
 long long faktorial(int n){
-	long long hasil ;
+	long hasil ;
 	hasil = 1;
 	for(bilangan = 1; bilangan <= n;bilangan++)
 	{
